refactor: extract BrowserStack upload helper
